+++ src/Runner.hs
@@ -1,0 +1,8 @@
+module Runner where
+
+import System.IO
+import Control.Monad
+import Evaluator
+
+getProgram :: String -> IO String
+getProgram file = readFile file